pyproject.toml
[project]
name = "multimodal-cancer-detection"
version = "0.1.0"
description = "Multimodal pancreatic cancer detection workflows, reports, and notebook-driven experiments."
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "numpy>=1.24.4,<1.26",
    "pandas>=1.5.3,<2.0",
    "PyYAML>=6.0,<7.0",
    "scikit-learn==1.3.2",
    "matplotlib>=3.7.2,<3.9",
    "seaborn>=0.12.2,<0.14",
    "tqdm>=4.66",
    "SimpleITK>=2.3",
    "nibabel>=5.1.0,<5.2",
    "pydicom>=2.4",
    "pillow>=10.0.0,<11.0.0",
    "opencv-python>=4.8,<5.0",
    "imbalanced-learn==0.11.0",
    "statsmodels>=0.14",
    "torch>=2.5,<2.7",
    "torchvision>=0.20,<0.22",
    "albumentations>=1.4,<2.1",
    "scikit-image>=0.22,<0.26",
    "pywinpty==2.0.13",
    "notebook==7.0.8",
    "ipykernel>=6.29",
    "pytest>=8.3,<9.0",
]

[tool.uv]
package = false

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
